% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{add_aav}
\alias{add_aav}
\title{Add AAV to the Projections Table}
\usage{
add_aav(projection_table, sources = c("RTS", "ESPN", "Yahoo", "NFL"))
}
\arguments{
\item{projection_table}{An output from the \link{projections_table} function}

\item{sources}{Which AAV sources should be added. should be one or more of
\code{c("RTS", "ESPN", "Yahoo", "NFL")}}
}
\description{
This function will add the AAV data to the projections table from the
\link{projections_table} function.
}
