% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{add_adp}
\alias{add_adp}
\title{Add ADP to the Projections Table}
\usage{
add_adp(projection_table, sources = c("RTS", "CBS", "ESPN", "Yahoo", "NFL",
  "FFC"))
}
\arguments{
\item{projection_table}{An output from the \link{projections_table} function}

\item{sources}{Which ADP sources should be added. should be one or more of
\code{c("RTS", "CBS", "ESPN", "Yahoo", "NFL", "FFC")}}
}
\description{
This function will add the ADP data to the projections table from the
\link{projections_table} function. It will add the average ADP from the sources
specfied, and the difference between the overall rank and ADP
}
