% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_be_deprecated.R
\name{add_risk}
\alias{add_risk}
\title{Add calculated risk to the table}
\usage{
add_risk(projection_table)
}
\description{
Calculation of risk is done by scaling the standard deviation variables
passed and averaging them before returning a measure with mean 5 and standard
deviation of 2
}
