% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{add_uncertainty}
\alias{add_uncertainty}
\title{Add uncertantity to the table}
\usage{
add_uncertainty(projection_table)
}
\description{
Calculation of uncertainty is done by scaling the standard deviation
variables, averaging them, and then creating a within-position percentile
rank ranging from 1 to 99. A score of 1 indicates there is very little
uncertainty (low standard deviation) and a score of 99 indicates there is
a large degree of uncertainty
}
\details{
A low score means there is general agreement among experts and projections.
A high score indicates there is a lot of variablility in rankings and/or
projections.
}
