% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_be_deprecated.R
\name{add_vor}
\alias{add_vor}
\title{Calculate VOR for Points, Ceiling and Floor}
\usage{
add_vor(tbl, vor_baseline = NULL)
}
\arguments{
\item{tbl}{The output from the \link{projected_points} function that has
been merged with the output from  he \link{confidence_interval} function}

\item{vor_baseline}{The VOR baseline values to be used. If omitted then the
\link{default_baseline} will be used}
}
\description{
This function calculates VOR for projected points as well as the floor and
ceiling values.
}
