% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_be_deprecated.R
\name{aggregate_stats}
\alias{aggregate_stats}
\title{Aggregate Projected Stats}
\usage{
aggregate_stats(data_result, src_weights = NULL)
}
\arguments{
\item{data_result}{An output from the \link{scrape_data} function.}

\item{src_weights}{A named vector with the weights from each source.
See \link{default_weights}}
}
\description{
This function aggregates the projected stats collected from each source with
the \link{scrape_data} function.
}
