% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrls.R
\name{analystOptions}
\alias{analystOptions}
\title{Analyst options for a period}
\usage{
analystOptions(period)
}
\arguments{
\item{scrapePeriod}{A \link{dataPeriod} object}
}
\description{
Find the analysts that are projecting stats for the provided period
}

