% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analystPositions}
\alias{analystPositions}
\title{Analyst positions}
\format{A \link[data.table]{data.table} with 151 rows and 4 columns
\describe{
 \item{analystId}{An integer identifying the analyst the row is referring to.
 Refers to the link{analyst} table.}
 \item{position}{Name of the position}
 \item{season}{Indicates whether the analysts projects seasonal data for the
 position.}
 \item{weekly}{Indicates whether the analysts projects weekly data for the
 position.}
}}
\usage{
analystPositions
}
\description{
Data table identifying which positions the different analysts are projecting
for
}
\keyword{datasets}

