% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analysts}
\alias{analysts}
\title{Analyst data}
\format{A \link[data.table]{data.table} with 27 rows and 7 columns
\describe{
 \item{analystId}{Unique integer identifier for the analyst}
 \item{analystName}{Name of analyst}
 \item{siteId}{An integer identifying the site that the analyst is projecting
 for. Refers to \link{sites} table.}
 \item{season}{Indicator of whether the analyst projects seasonal data or not}
 \item{weekly}{Indicator of whether the analyst projects weekly data or not}
 \item{sourceId}{The identifier that the site uses for the analyst if any}
 \item{weight}{The weight used for the analyst in the weighted calculations}
}}
\usage{
analysts
}
\description{
Data table with information on analysts that are bing used for data scrapes.
}
\keyword{datasets}

