% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationDataPrep.R
\name{avgValue}
\alias{avgValue}
\title{Calculate average based on selected method}
\usage{
avgValue(calcMethod = "weighted", dataValue = as.numeric(),
  dataWeights = as.numeric(), na.rm = FALSE)
}
\arguments{
\item{calcMethod}{One of \code{c("average", "weighted", "robust")}}

\item{dataValue}{A numeric vector of values to average base on chosen calculation
method.}

\item{dataWeights}{A numeric vector of weight values associated with the
dataValue parameter.}

\item{na.rm}{A logical value determining if NA values should be removed.}
}
\description{
Will calculate average of provided data and weights based on the selected
method
}

