% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStdDev.R
\name{calcStdDev}
\alias{calcStdDev}
\title{Calculate Standard Deviation}
\usage{
calcStdDev(calcMethod = "weighted", dataValue = as.numeric(),
  dataWeights = as.numeric(), na.rm = FALSE)
}
\arguments{
\item{calcMethod}{Calculation method. Can be one of "weighted", "robust" or
"average"}

\item{dataValue}{Vector of values to calculate standard deviation for}

\item{dataWeights}{Vector of weights for weighted calculation}
}
\description{
Standard Deviation is calculated based on method passed to the function.
}

