% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_be_deprecated.R
\name{calculate_risk}
\alias{calculate_risk}
\title{Risk calculation based on two variables}
\usage{
calculate_risk(var1, var2)
}
\description{
Calculation of risk is done by scaling the standard deviation variables
passed and averaging them before returning a measure with mean 5 and standard
deviation of 2
}
