% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{calculate_uncertainty}
\alias{calculate_uncertainty}
\title{Uncertainty calculation}
\usage{
calculate_uncertainty(..., percentage = TRUE)
}
\description{
Calculation of uncertainty returns a value from 1 to 99 where higher values
indicate more uncertainty (i.e., more variability).
}
