% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePoints.R
\name{calculatePoints}
\alias{calculatePoints}
\title{Caluclate fantasy points}
\usage{
calculatePoints(projectionData = data.table(), scoringRules = list())
}
\arguments{
\item{projectionData}{A \link[data.table]{data.table} with projected stats.
Should from the \link{getMeltedData} function}

\item{scoringRules}{A list of scoring rules with one element per position}
}
\value{
A \link[data.table]{data.table} with up to 4 cloumns
\describe{
 \item{position}{The position the player is playing}
 \item{playerId}{The ID of the player. Get player names by merging results from
 \link{getPlayerData}}
 \item{analyst}{The ID of the analyst which projections are used as the basis
 for the points}
 \item{points}{The calculated number of fantasy points}
}
}
\description{
Based on list of scoring rules calculate fantasy points using projection data
passed in a table
}

