% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRisk.R
\name{calculateRisk}
\alias{calculateRisk}
\title{Risk calculation based on to variables}
\usage{
calculateRisk(var1, var2)
}
\description{
Calculation of risk is done by scaling the standard deviation variables
passed and averaging them before returning a measure with mean 5 and standard
deviation of 2
}

