% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectPoints.R
\name{calculateVor}
\alias{calculateVor}
\title{Calculate Value Over Replacement}
\usage{
calculateVor(ranks, points, position)
}
\arguments{
\item{ranks}{Player ranks}

\item{points}{Player Points}

\item{position}{Player Position. Used to extract value from \link{vorBaseline}
and \link{vorAdjustment}}
}
\description{
Based on provided ranks and points calculate the value over replacement.
Function uses \link{vorBaseline} and \link{vorAdjustment} in the calculation.
Please adjust these to match your league before running calculation.
}

