% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp_functions.R
\name{cbs_draft}
\alias{cbs_draft}
\title{Get ADP data from CBS}
\usage{
cbs_draft(metric = "adp")
}
\value{
A \link{data.frame} with the results.
}
\description{
This function scrapes ADP data from CBS Sports
}
