% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching_helpers.R
\name{clear_ffanalytics_cache}
\alias{clear_ffanalytics_cache}
\title{Manually clear cache associated with ffanalytics package}
\usage{
clear_ffanalytics_cache(ffa_objects = NULL)
}
\arguments{
\item{ffa_objects}{(\emph{optional}) a character vector of object names (i.e.,
returned by \code{list_ffanalytics_cache()})}
}
\description{
Clears all scraped data cache associated with the ffanalytics package, or a
subset of objects from the cache
}
