% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_be_deprecated.R
\name{confidence_interval}
\alias{confidence_interval}
\title{Calculate the Upper and Lower Limits for Projected Points}
\usage{
confidence_interval(src_pts, weights = NULL)
}
\arguments{
\item{src_pts}{An output from the \link{source_points} function}

\item{weights}{A named vector with the weights from each source.
See \link{default_weights}}
}
\description{
This function calculates the ceiling and floor for projected points from
different sources based on quantiles
}
