% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceInterval.R
\name{confidenceInterval}
\alias{confidenceInterval}
\title{Calculate confidence interval}
\usage{
confidenceInterval(calcMethod = "weighted", dataValue = as.numeric(),
  dataWeights = as.numeric(), pValues = c(0.1, 0.9), na.rm = FALSE)
}
\arguments{
\item{calcMethod}{Calculation method. Can be one of "weighted", "robust" or
"average"}

\item{dataValue}{Vector of values to calculate confidence interval for}

\item{dataWeights}{Vector of weights for weighted calculation}

\item{pValues}{Vector of percentiles for the confidence interval. Defaults to
\code{c(0.1, 0.9)} for 10 and 90 percentiles.}
}
\description{
Confidence intervals are determined as percentiles (default 10 and 90).
If the correcpsonding average value is less than the low point then we
set the lower point to the minimum value. Conversely if the average value is
higher than the upper limit then we set the upper limit to the max value.
}

