% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalConstants.R
\docType{data}
\name{Constants}
\alias{Constants}
\alias{ffnAPI}
\alias{position.Id}
\alias{position.name}
\alias{yahooLeague}
\title{Package Constants}
\usage{
position.name

position.Id

yahooLeague

ffnAPI
}
\description{
Constants build into the package
}
\details{
A number of constants are included in the package
}
\keyword{datasets}

