% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createObject.R
\name{createObject}
\alias{createObject}
\title{Create objects from list}
\usage{
createObject(objName = as.character(), targetValues = list())
}
\arguments{
\item{objName}{Name of object to be created}

\item{targetValues}{List of values representing slots in the object}
}
\description{
Allows creation of class objects based on a list.
}

