% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataResult.R
\docType{class}
\name{dataResult-class}
\alias{dataResult}
\alias{dataResult-class}
\title{Class to represent scrape results}
\description{
Class to represent scrape results
}
\section{Slots}{

\describe{
\item{\code{position}}{Position that the scrape data isfor}

\item{\code{resultData}}{A \link{data.table} holding the data. The columns in the
table are determined by the \link{staticColumns} and \link{resultColumns}.}
}}

