% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\docType{data}
\name{default_baseline}
\alias{default_baseline}
\title{Default VOR Baseline}
\format{An object of class \code{numeric} of length 9.}
\usage{
default_baseline
}
\description{
This is the default baseline that is used if not otherwise specified when
calculating VOR:
\code{c(QB = 13, RB = 35, WR = 36, TE = 13, K = 8, DST = 3, DL = 10, LB = 10, DB = 10)}
}
\keyword{datasets}
