% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\docType{data}
\name{default_threshold}
\alias{default_threshold}
\title{Default Threshold Values for Tiers}
\format{An object of class \code{numeric} of length 9.}
\usage{
default_threshold
}
\description{
These are the default threshold values used when applying Cohen's D values
to determine tiers:
\code{c(QB = 1, RB = 1, WR = 1, TE = 1, K = 1, DST = 0.1, DL = 1, DB = 1, LB = 1)}
}
\keyword{datasets}
