% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\docType{data}
\name{default_weights}
\alias{default_weights}
\title{Default Weights for Projection Sources}
\format{
An object of class \code{numeric} of length 14.
}
\usage{
default_weights
}
\description{
These are the weights that are used for each source when calculation weighted
averages and standard deviations if no weights are specified.
\code{c(CBS = 0.145, Yahoo = 0.000,  ESPN = 0.157,  NFL = 0.140,
FFToday = 0.151, NumberFire = 0.142, FantasyPros = 0.000,
FantasySharks= 0.142, FantasyFootballNerd = 0.000,
WalterFootball = 0.130, RTSports= 0.123,
FantasyData= 0.000, FleaFlicker = 0.000, FanDuel = 0.142)}
}
\keyword{datasets}
