% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\docType{data}
\name{default_weights_by_src}
\alias{default_weights_by_src}
\title{Default Weights for Projection Sources}
\format{
An object of class \code{list} of length 9.
}
\usage{
default_weights_by_src
}
\description{
Object with default weights for each position by source
}
\keyword{datasets}
