% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp_functions.R
\name{espn_draft}
\alias{espn_draft}
\title{Get ADP/AAV data from ESPN}
\usage{
espn_draft()
}
\value{
A \link{data.frame} with the results. Contains both ADP and AAV
}
\description{
This function scrapes ADP and AAV data from ESPN
}
