% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagueScoring.R
\docType{data}
\name{ffa.vorAdjustment}
\alias{ffa.vorAdjustment}
\title{Default VOR Adjustments}
\format{An object of class \code{numeric} of length 9.}
\usage{
ffa.vorAdjustment
}
\description{
Default VOR Adjustments
}
\keyword{datasets}

