% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagueScoring.R
\docType{data}
\name{ffa.vorBaseline}
\alias{ffa.vorBaseline}
\title{Default VOR Baseline}
\format{An object of class \code{numeric} of length 9.}
\usage{
ffa.vorBaseline
}
\description{
Default VOR Baseline
}
\keyword{datasets}

