% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp_functions.R
\name{ffc_draft}
\alias{ffc_draft}
\title{Get ADP data from Fantasy Football Calculator}
\usage{
ffc_draft(
  format = c("standard", "ppr", "2qb", "dynasty", "rookie"),
  pos = c("all", "qb", "rb", "wr", "te", "def", "pk"),
  season = 2019
)
}
\arguments{
\item{format}{Indicates which league format data should be returned for.
Should be one of \code{c("standard", "ppr", "2qb", "dynasty", "rookie")}}

\item{pos}{indicates the position the data should be returned for. Should be
one of \code{c("all", "qb", "rb", "wr", "te", "def", "pk")}}

\item{season}{Indicates the season that data should be retuned for.}
}
\value{
A \link{data.frame} with the results.
}
\description{
This function scrapes ADP data from Fantasy Football Calculator
}
