% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{firstLast}
\alias{firstLast}
\title{Reverse Last and First Name}
\usage{
firstLast(lastFirst)
}
\arguments{
\item{lastFist}{A string with the name to be converted}
}
\description{
Takes an input string in the form of "Last Name, First Name" and converts it
to "First Name Last Name".
}
\examples{
firstLast("Smith, John") # Will return John Smith
}

