% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp_functions.R
\name{get_adp}
\alias{get_adp}
\title{Get ADP/AAV data from multple sources}
\usage{
get_adp(
  sources = c("RTS", "CBS", "Yahoo", "NFL", "FFC", "MFL", "ESPN"),
  metric = c("adp", "aav")
)
}
\arguments{
\item{sources}{Indicates what sources to retrieve data from. Choose from
\code{c("RTS", "CBS", "ESPN", "Yahoo", "NFL")}. Multiple sources are allowed.
If omitted all sources will be scraped.}

\item{type}{Specifies what data to collect. Should one of \code{c("ADP", "AAV")}.
If omitted then ADP data will be scraped.}
}
\value{
A \link{data.frame} with the results. The player's id from the
\code{player_ids} table and a column for each source. The average value is also
returned if multiple sources are specified
}
\description{
This function scrapes ADP or AAV data from multiple sources
}
