% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdpData.R
\name{getADPdata}
\alias{getADPdata}
\title{Get ADP data}
\usage{
getADPdata(ADPsources = c("CBS", "ESPN", "FFC", "MFL", "NFL"),
  season = as.POSIXlt(Sys.Date())$year + 1900, teams = 12,
  format = "standard", mflMocks = NULL, mflLeagues = NULL)
}
\arguments{
\item{season}{The season the ADP data is from}

\item{teams}{Number of teams in the league}

\item{format}{The format of the league, i.e. standard, ppr}

\item{mflMocks}{Include mock drafts from MFL. Set to 1 if only mock drafts
should be used, 0 if only real drafts should be used. If not speficied all
types of drafts will be used.}

\item{mflLeagues}{What type of leagues to include for MFL. Set to 0 to use
redraft leagues only; 1 to only use keeper leagues, 2 for rookie drafts, and
3 for MFL Public Leagues. If not speficied all types of drafts will be used.}

\item{ADPsource}{Character vector with one or more of
\code{"CBS", "ESPN", "FFC", "MFL", "NFL"}}
}
\description{
Retrieve ADP data from multiple sources and combine into average.
}

