% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdpAav.R
\name{getCBSValues}
\alias{getCBSValues}
\title{ADP data from CBS}
\usage{
getCBSValues()
}
\value{
\link[data.table]{data.table} with 3 columns:
\describe{
 \item{cbsId}{Player ID from CBS. Merge with results from \link{getPlayerData}
 to get player names}
 \item{adp}{Average ADP}
 \item{leagueType}{Assuming standard league type}}
}
\description{
Retrieve ADP data from CBS
}

