% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdpAav.R
\name{getESPNValues}
\alias{getESPNValues}
\title{ADP and auction value data from ESPN}
\usage{
getESPNValues()
}
\value{
\link[data.table]{data.table} with 6 columns:
\describe{
 \item{player}{Name of player}
 \item{position}{Player position}
 \item{team}{Team the player is playing for}
 \item{adp}{Average ADP}
 \item{aav}{Average auction value}
 \item{leagueType}{Assuming standard league type}}
}
\description{
Retrieve ADP and auction value from ESPN
}

