% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdpAav.R
\name{getFFCValues}
\alias{getFFCValues}
\title{ADP data from FantasyFootballCalculator.com}
\usage{
getFFCValues(format = "standard", teams = 12)
}
\arguments{
\item{format}{Format of league. Can be one of \code{"standard", "ppr", "2qb", "dynasty", "rookie"}.}

\item{teams}{Numer of teams in the league. One of 8, 10, 12, 14}
}
\value{
\link[data.table]{data.table} with 5 columns:
\describe{
 \item{player}{Name of player}
 \item{position}{Player position}
 \item{team}{Team the player is playing for}
 \item{adp}{Average ADP}
 \item{leagueType}{Chosen format for the league}}
}
\description{
Retrieve ADP data from fantasyfootballcalculator.com.
}

