% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationDataPrep.R
\name{getMeltedData}
\alias{getMeltedData}
\title{Melt data into long form}
\usage{
getMeltedData(dataResult)
}
\arguments{
\item{dataResult}{A \link{dataResult} object from the data scrape}
}
\description{
Takes the data result with a projected stat in each column and melts the
data into one row per player per stat column.
}

