% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdpAav.R
\name{getMFLValues}
\alias{getMFLValues}
\title{ADP and auction value data from MyFantasyLeague.com}
\usage{
getMFLValues(season = as.POSIXlt(Sys.Date())$year + 1900, type = "adp",
  teams = -1, ppr = -1, mock = -1, keeper = -1)
}
\arguments{
\item{season}{Year the data is retrieved for}

\item{type}{One of "adp" or "aav" to indicate whether ADP or auction values
should be retrieved.}

\item{teams}{Number of teams. If specified only drafts with that number of
teams will be include}

\item{ppr}{Specify if only ppr or non-ppr drafts should be considered. Set to 1
if only ppr drafts should be used, 0 if only standard drafts should be used.
If not specified all types of draft will be used.}

\item{mock}{Specify if only mock or real drafts should be used. Set to 1 if
only mock drafts should be used, 0 if only real drafts should be used. If not
speficied all types of drafts will be used.}

\item{keeper}{Specify to select what types of leagues should be used. Set to 0
to use redraft leagues only; 1 to only use keeper leagues, 2 for rookie drafts,
and 3 for MFL Public Leagues. If not  speficied all types of drafts will be used.}
}
\value{
\link[data.table]{data.table} wih up to 5 columns:
\describe{
 \item{mflId}{Player ID from MFL Merge with results from \link{getPlayerData}
 to get player names}
 \item{selectedIn}{Number of drafts player has been selected in}
 \item{aav}{Average auction value}
 \item{adp}{ADP}
 \item{minPick}{Earliest pick}
 \item{maxPick}{Latest pick}
 }
}
\description{
Retrieve ADP and auction value data from MyFantasyLeague.com
}

