% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdpAav.R
\name{getNFLValues}
\alias{getNFLValues}
\title{ADP data from FantasyFootballCalculator.com}
\usage{
getNFLValues()
}
\value{
\link[data.table]{data.table} with 5 columns:
\describe{
 \item{esbid}{Player ID for the player. Merge with results from \link{getPlayerData}
 to get player names}
 \item{player}{Name of player}
 \item{position}{Player position}
 \item{adp}{Average ADP}
 \item{aav}{Average auction value}
 }
}
\description{
Retrieve ADP data from fantasyfootballcalculator.com.
}

