% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerData.R
\name{getPlayerData}
\alias{getPlayerData}
\title{Combine MFL and NFL player data}
\usage{
getPlayerData(season, weekNo, pos = position.name)
}
\arguments{
\item{season}{The year that player data is to be retrieved from}

\item{weekNo}{The weekNo that the player data is to be retrieved from}

\item{pos}{A character vector with position names to be retrieved}
}
\value{
A \link[data.table]{data.table} with 13 columns
\describe{
 \item{playerId}{NFL's ID for the player}
 \item{player}{Name of player}
 \item{yahooId}{Yahoo's ID for the player}
 \item{cbsId}{CBS's ID for the player}
 \item{mflId}{MFL's ID for the player}
 \item{position}{The position the player is playing}
 \item{team}{The team the player is playing for}
 \item{draftYear}{The year the player was drafted}
 \item{birthData}{The players birth date}
 \item{rookie}{A logical value indicating whether the player is a rookie}
 \item{opponent}{Team the player is facing next}
 \item{depthChart}{Number on the depth chart for the player 1 = starter}
 \item{esbid}{Alternate ID for player. Used for ADP/AAV data from NFL}
}
}
\description{
Retrieve data from MFL and NFL and combine
}

