% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerName.R
\name{getPlayerName}
\alias{getPlayerName}
\title{Clean Player Data For Projections}
\usage{
getPlayerName(playerCol)
}
\arguments{
\item{playerCol}{The vector of player data taken from the data table that is returned from
the data scrape}
}
\value{
The updated vector of player data
}
\description{
For many of the the data soruces the player column contains more data than
needed to identify the player. With the help of regular expression data such as position,
team and injury information is cleaned from the player names.
}

