% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanks.R
\name{getRanks}
\alias{getRanks}
\title{Expert Consensus Rankings}
\usage{
getRanks(rank.position = "consensus", leagueType = "std", weekNo = 0)
}
\arguments{
\item{rank.position}{Position to retrieve ranks for. Use "consensus" to get
overall rankings}

\item{leagueType}{Indicate whether to get ppr rankings or standard rankings (std)}

\item{weekNo}{Week number to retrieve ranks for. Use 0 for season ranks.}
}
\value{
\link[data.table]{data.table} with up to 11 columns:
\describe{
 \item{player}{Name of player}
 \item{position}{Player's position}
 \item{team}{Team the player is playing for}
 \item{ecrRank}{Consensus Rank}
 \item{bestRank}{Highest rank from experts}
 \item{worstRank}{Lowest rank from experts}
 \item{avgRank}{Average rank from experts}
 \item{sdRank}{Standard deviation of ranks}
 \item{adp}{Average Draft Position}
 \item{vsAdp}{Difference between overall rank and ADP}
 \item{rankType}{Will have value of "overall" or "position"}
}
}
\description{
Reterieve expert consensus rankings from fantasypros.com
}

