% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrls.R
\name{getUrls}
\alias{getUrls}
\title{URLs to scrape}
\usage{
getUrls(selectAnalysts = analysts$analystId,
  period = periodType(dataPeriod()), positions = position.name)
}
\arguments{
\item{selectAnalysts}{An integer vector of selected analystIds.
See \link{analysts} for possible values}

\item{period}{A string indicating wether the URLs to produce are for weekly
or seasonal data scrape}

\item{positions}{A character vector of positions to scrape.}
}
\description{
Based on selected analysts, a given period and selected analysts generate
URLs that will be scraped. URLs are generated as merges of the \link{analysts},
\link{sites}, \link{siteUrls}, and \link{siteTables} datasets.
}

