% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdpAav.R
\name{getYahooValues}
\alias{getYahooValues}
\title{ADP and auction value data from Yahoo}
\usage{
getYahooValues(type = "SD")
}
\arguments{
\item{type}{Draft type: AD for auction draft, SD for standard draft.}
}
\value{
\link[data.table]{data.table} with 3 columns:
\describe{
 \item{cbsId}{Player ID from CBS. Merge with results from \link{getPlayerData}
 to get player names}
 \item{adp/aav}{Average ADP if type = "SD"; Average auction value if type = "AD"}
 \item{leagueType}{Assuming standard league type}}
}
\description{
Retrieve ADP and auction value data from Yahoo
}

