% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_classes.R
\docType{class}
\name{html_source}
\alias{html_source}
\title{HTML source object}
\format{An R6 class object.}
\usage{
html_source
}
\description{
This objects represent the basic set of data needed for defining a projection
source with HTML data. It is an extesnsion of the \link{projection_source}
object with some specific fields and methods related to scraping HTML data.
}
\section{Fields}{

\describe{
\item{\code{table_css}}{A string with a CSS selector identifying the HTML
\code{<table>} element holding the projection data.}

\item{\code{pid_css}}{A string with a CSS selector idenfitying the HTML node holding
the player id if available}

\item{\code{rm_elem}}{A character vector of CSS selectors identifying HTML nodes to
remove for successful scraping of the table.}

\item{\code{index}}{If \code{table_css} does not uniquely identify the table, use this
field to identify the index number for the table in the list of nodes. If
multiple numbers are specified then the tables are ssume to have identical
number of rows and will be combined with \link{bind_cols}}

\item{\code{extract_pid}}{A function that will take the HTML node holding the player
id and extract the specific player_id}

\item{\code{split_cols}}{A list with each element being a list representing input to
either \link{separate} or \link{extract}. Each input element should be in the
format of a function of position to allow for different handling of fields for
different positions. See the \code{projection_sources} object for predefined
sources.}

\item{\code{recode_cols}}{a list with each element being a list representing names
of columns to be recoded and a named vector for each column holding the recode
values to be used by \link{recode}.}
}}

\note{
See \url{https://www.w3schools.com/cssref/css_selectors.asp} for details
on defining CSS selectors
}
\section{Methods}{

\describe{
  \item{\code{open_session}}{Takes \code{season, week, position} as input and
  and opens a session on the website via the \link{html_session} function after
  determining the URL}
  \item{\code{close_session}}{Closes the session that is currently open}
  \item{\code{get_table}}{Retrieves the table from the session without any
  spliltting of columns defined in \code{split_cols}, any recoding as defined
  in \code{recode_cols} or any renaming based on \code{stat_cols}}
  \item{\code{scrape}}{Scrapes data from the table specified and wrangles the
  columns based on \code{split_cols}, \code{recode_cols} and \code{stat_cols}}
}
}

\seealso{
Other source-class: \code{\link{json_source}},
  \code{\link{projection_source}},
  \code{\link{xlsx_source}}
}
\keyword{internal}
