% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_funcs.R
\name{Impute}
\alias{Impute}
\alias{impute_na_off}
\alias{kick_impute}
\title{Imputing values for missing data}
\usage{
impute_na_off(tbl)

kick_impute(kick_tbl)
}
\description{
Differnt sources provide different data. The functions below will fill out
NA values for some sources with values based on data from other sources
}
