% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching_helpers.R
\name{list_ffanalytics_cache}
\alias{list_ffanalytics_cache}
\title{Checks the scrapes that are currently cached}
\usage{
list_ffanalytics_cache(quiet = FALSE)
}
\arguments{
\item{quiet}{whether the function should return a message if cache is empty}
}
\value{
A \code{data.frame} with the object (scrape) name, and time since caching
}
\description{
Checks the cached data, removes cached objects >= 8 hours old, and returns a list
of objects that are currently cached
}
