% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{mad2}
\alias{mad2}
\title{Altered MAD function}
\usage{
mad2(
  x,
  center = median(x),
  constant = 1.4826,
  na.rm = FALSE,
  low = FALSE,
  high = FALSE,
  w
)
}
\description{
NA for length = 1 & an ignored weight function
}
