% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_scoring.R
\name{make_scoring_tables}
\alias{make_scoring_tables}
\title{Create Scoring Tables when computing projections table}
\usage{
make_scoring_tables(scoring_rules)
}
\description{
Creates tables for the DST points brackets and the scoring tables by position
used in the `projections_table()` function.
}
