% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp_functions.R
\name{mfl_draft}
\alias{mfl_draft}
\title{Get ADP or AAV data from MyFantasyLeague}
\usage{
mfl_draft(
  metric = c("adp", "aav"),
  period = c("RECENT", "ALL", "DRAFT", "JUNE", "JULY", "AUG1", "AUG15", "START", "MID",
    "PLAYOFF"),
  format = c("All Leagues", "PPR", "Std"),
  nteams = c(12, 8, 10, 14, 16),
  is_keeper = c("No", "Keeper", "Rookie Only"),
  is_mock = c("No", "Mock", "All Leagues"),
  cutoff = 10
)
}
\arguments{
\item{metric}{Indicated whether to pull ADP (default) or AAV}

\item{period}{Includes metric for drafts following this time-period}

\item{format}{Scoring system for receptions}

\item{nteams}{Number of teams in the league}

\item{is_keeper}{Whether or not the league is a keeper league and/or a rookie draft}

\item{is_mock}{Whether or not it is a mock draft}

\item{cutoff}{Includes players in at least __\% of drafts. Default to 10\% of drafts.}
}
\value{
A \link{data.frame} with the results.
}
\description{
This function scrapes ADP or AAV data from MyFantasyLeague. More details on
the API available at \link{https://api.myfantasyleague.com/2023/api_info?STATE=details}
}
