% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mflPlayers.R
\name{mflPlayers}
\alias{mflPlayers}
\title{Read MFL Player Data}
\usage{
mflPlayers(season = 2016, weekNo = 0, pos = position.name)
}
\arguments{
\item{season}{The year that player data is to be retrieved from}

\item{weekNo}{The weekNo that the player data is to be retrieved from}

\item{pos}{A character vector with position names to be retrieved}
}
\value{
A \link[data.table]{data.table} with 10 columns
\describe{
 \item{playerId}{NFL's ID for the player}
 \item{player}{Name of player}
 \item{yahooId}{Yahoo's ID for the player}
 \item{cbsId}{CBS's ID for the player}
 \item{mflId}{MFL's ID for the player}
 \item{position}{The position the player is playing}
 \item{team}{The team the player is playing for}
 \item{draftYear}{The year the player was drafted}
 \item{birthData}{The players birth date}
 \item{rookie}{A logical value indicating whether the player is a rookie}
}
}
\description{
Function to read player data from MFL using the MFL API
}

