% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp_functions.R
\name{nfl_draft}
\alias{nfl_draft}
\title{Get ADP/AAV data from NFL}
\usage{
nfl_draft()
}
\value{
A \link{data.frame} with the results. Contains both ADP and AAV data
}
\description{
This function scrapes ADP or AAV data from NFL
}
