% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflPlayerData.R
\name{nflPlayerData}
\alias{nflPlayerData}
\title{Player Data from NFL.com}
\usage{
nflPlayerData(season = 2016, weekNo = 0, positions = position.name)
}
\arguments{
\item{season}{The year data is to be retrieved from}

\item{weekNo}{The week that data is to be retrieved from}

\item{positions}{A character vector of positions to be retrieved}
}
\value{
A \link[data.table]{data.table} with 7 columns:
\describe{
 \item{playerId}{NFL's ID For the player}
 \item{player}{Name of player}
 \item{position}{NFL's position designation for the player}
 \item{team}{NFL Team that the player is playing for}
 \item{opponent}{Team the player is facing next}
 \item{depthChart}{Number on the depth chart for the player 1 = starter}
 \item{esbid}{Alternate ID for player. Used for ADP/AAV data from NFL}
}
}
\description{
Retrieve player data from NFL.com
}

