% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPeriod.R
\docType{methods}
\name{periodType,dataPeriod-method}
\alias{periodType,dataPeriod-method}
\title{Determine if period is a week or season}
\usage{
\S4method{periodType}{dataPeriod}(object)
}
\arguments{
\item{x}{A dataPeriod object}
}
\description{
Determine if period is a week or season
}

