% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_be_deprecated.R
\name{points_sd}
\alias{points_sd}
\title{Calculate Standard Deviations for Projected Points}
\usage{
points_sd(src_pts, weights = NULL)
}
\arguments{
\item{src_pts}{An output from the \link{source_points} function}

\item{weights}{A named vector with the weights from each source.
See \link{default_weights}}
}
\description{
This function calculates the standard deviaion for projected points from
different sources
}
