% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{projected_points}
\alias{projected_points}
\title{Calculate Projected Points}
\usage{
projected_points(agg_stats, scoring_rules)
}
\arguments{
\item{agg_stats}{An output from the \link{aggregate_stats} function}

\item{scoring_rules}{The scoring rules to be used.}
}
\description{
This function calculates the projected points for each player based on the
aggregated stats from the \link{aggregate_stats} function. The resulting table
contains the projected points, the position rank and the points drop-off for
each player.
}
