% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{projections_table}
\alias{projections_table}
\title{Create a Projections Table}
\usage{
projections_table(
  data_result,
  scoring_rules = NULL,
  src_weights = NULL,
  vor_baseline = NULL,
  tier_thresholds = NULL
)
}
\arguments{
\item{data_result}{An output from the \link{scrape_data} function}

\item{scoring_rules}{The scoring rules to be used for calculations. See
\code{vignette("scoring_settings")} on how to define custom scoring settings.
If omitted then default \link{scoring} settings will be used.}

\item{src_weights}{A named vector defining the weights for each source to be
used in calculations. If omitted then \link{default_weights} will be used.}

\item{vor_baseline}{A named vector defineing the baseline to use for VOR
calculations. If omitted then the \link{default_baseline} will be used.}

\item{tier_thresholds}{The threshold values to be used when determining tiers.
If omitted then the \link{default_threshold} will be used.}
}
\description{
This function creates the projections table based on the scraped data from the
\link{scrape_data} function. The output is a table containing the projected
points, confidence intervals, standard deviation for points, and if seasonal
data also the VOR values
}
