% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{projections_table2}
\alias{projections_table2}
\title{New, lighter, projections_table function}
\usage{
projections_table2(
  data_result,
  scoring_rules = NULL,
  src_weights = NULL,
  vor_baseline = NULL,
  tier_thresholds = NULL,
  avg_type = c("average", "robust", "weighted"),
  return_raw_stats = FALSE
)
}
\description{
Testing & improving now TO replace the current function in the next major app update
}
