% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectPoints.R
\name{projectPoints}
\alias{projectPoints}
\title{Calculate points}
\usage{
projectPoints(projectionData, scoringRules, avgType = "average")
}
\arguments{
\item{projectionData}{A \link[data.table]{data.table} with projected stats}

\item{scoringRules}{A list with tables for league scoring rules. See
\link{scoringRules} for reference on format}

\item{avgType}{Which average to use. Should be one of \code{average, robust, weighted}}
}
\description{
For the scraped data, projected points, confidence interval, standard deviation
and position ranks are calculated
}

