% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUrl.R
\name{readUrl}
\alias{readUrl}
\title{Read data from a URL}
\usage{
readUrl(inpUrl, columnTypes, columnNames, whichTable, removeRow, dataType,
  idVar, playerLinkString)
}
\arguments{
\item{inpUrl}{The URL to get data from}

\item{columnTypes}{A character vector describing the types of columns in the
data. Note: \code{length(columnTypes) == length(columnNames)}.}

\item{columnNames}{A character vector describing the names of the columns in
the data. Note: \code{length(columnTypes) == length(columnNames)}.}

\item{whichTable}{A number or character describing the table to get. This
can be leveraged for HTML tables and spreadsheet files.}

\item{removeRow}{A numeric vector indicating rows to skip at the top of the data.
For exampe \code{c(1,2)} will skip the first two rows of data.}

\item{dataType}{A character indicating the type of data (HTML, XML, file, xls)}
}
\value{
Returns a \link{data.table} with data from URL.
}
\description{
The task for this function is to read the data from the URL location and assign
appropriate column names. The function will throw a warning if there are
more columns in the data table than expected. If there are fewer columns than
expected the function will retry up to 10 times to get the number of columns
correct. If it fails after 10 tries then an error will be thrown.
}

