% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationDataPrep.R
\name{redistributeValues}
\alias{redistributeValues}
\title{Redistribute values}
\usage{
redistributeValues(valueTable = data.table(), calcType = "weighted",
  fromVar = "fg", toVars = c("fg0019", "fg2029", "fg3039", "fg4049",
  "fg50"), excludeAnalyst = 20)
}
\arguments{
\item{valueTable}{A \link[data.table]{data.table}. Assumes outout from the
\link{getMeltedData} function.}

\item{calcType}{A string specifying which calculation method to use for the
average values}

\item{fromVar}{A string specifying the name of the variable to distribute from}

\item{toVars}{A character vector with the names of variables to distribute to}

\item{excludeAnalyst}{An integer indicating an analyst to exclude. This will
exclude the analyst from the averages}
}
\description{
Allows for the redistribution of values from \bold{one} variable to a set of
others based on the averages. For example, the function can be used to
redistribute total field goals to field goals per distances based of what
the average values are for each of the field goals per distance.
}

