% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationDataPrep.R
\name{replaceMissingData}
\alias{replaceMissingData}
\title{Find replacement data for missing values}
\usage{
replaceMissingData(statData = data.table(), calcType = "weighted")
}
\arguments{
\item{statData}{A \link[data.table]{data.table}. Assumes outout from the
\link{getMeltedData} function.}

\item{calcType}{A string specifying which calculation method to use for the
average values}
}
\description{
For analysts that don't report on certain values the averages across other
analysts are calculated so they can be imputed.
}

