% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{retrieveData}
\alias{retrieveData}
\title{Retrive data from a source table}
\usage{
retrieveData(srcTbl, srcPeriod, fbgUser = NULL, fbgPwd = NULL)
}
\arguments{
\item{srcTbl}{A \link{sourceTable} object representing the table to get data from}

\item{srcPeriod}{A \link{dataPeriod} object representing the period to get}

\item{fbgUser}{User Name for an active footballguys.com account. Needed if
data scrape is requested from Footballguys}

\item{fbgPwd}{Password for an active footballguys.com account. Needed if
data scrape is requested from Footballguys}

\item{srcAnalyst}{A \link{sourceAnalyst} object representing the analyst
to scrape data from}
}
\value{
scrapeData a scrapeData object
}
\description{
Data can be retrieved from a source table when specified along with a source
analyst and a data period. The data scrape will translate the data columns
for each source table into a uniform format.
}

