% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp_functions.R
\name{rts_draft}
\alias{rts_draft}
\title{Get ADP/AAV data from RTSports}
\usage{
rts_draft(aav = FALSE)
}
\arguments{
\item{aav}{Indicates whether AAV data is scraped. If set to \code{FALSE} (Default),
ADP data is scraped. Set it to \code{TRUE} to scrape AAV data.}
}
\value{
A \link{data.frame} with the results.
}
\description{
This function scrapes ADP or AAV data from RTSports
}
