% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScrape.R
\name{runScrape}
\alias{runScrape}
\title{Scrape Projections}
\usage{
runScrape(season = NULL, week = NULL, analysts = NULL, positions = NULL,
  fbgUser = NULL, fbgPwd, updatePlayers = TRUE)
}
\arguments{
\item{season}{The season of projections to scrape (e.g., 2015).}

\item{week}{The week number of projections to scrape (e.g., 16).
Week number should be an integer between 0 and 21.
Week number 0 reflects seasonal projections.
Week number between 1 and 17 reflects regular season projections.
Week number between 18 and 21 reflects playoff projections.}

\item{analysts}{An integer vector of analystIds specifying which analysts' projections to
scrape. See \link{analysts} data set for values of analystIds.}

\item{positions}{A character vector of position names specifying which positions
to scrape: \code{c("QB", "RB", "WR", "TE", "K", "DST", "DL", "LB", "DB")}.}
}
\value{
list of \link{dataResults}. One entry per position scraped.
}
\description{
Executes a scrape of players' fantasy football projections based on the selected
season, week, analysts, and positions. If no inputs are specified, the user is prompted.
}
\note{
The function has the ability to include subscription based sources,
but you will need to either download subscription projections separately or
provide a user name and password for those sites.
Scraping past seasons/weeks is nearly impossible because very few if any sites
make their historical projections available. An attempt to scrape historical
projections will likely produce current projections in most cases.
}
\examples{
runScrape(season = 2016, week = 0,         ## Scrape 2016 season data for all
         analysts = 99, positions = "all") ## available analysts and positions

runScrape(season = 2016, week = 1,               ## Scrape 2016 week 1 data for
         analysts = c(-1, 5),                    ## CBS Average and NFL.com
         positions = c("QB", "RB", "WR", "TE",)) ## and offensive positions
}

