% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagueScoring.R
\docType{data}
\name{scoreThreshold}
\alias{scoreThreshold}
\title{Default Scoring threshold for tiers}
\format{An object of class \code{numeric} of length 9.}
\usage{
scoreThreshold
}
\description{
Default Scoring threshold for tiers
}
\keyword{datasets}

