% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_rules.R
\docType{data}
\name{scoring}
\alias{scoring}
\title{Default scoring rules for the calculations}
\format{
An object of class \code{list} of length 9.
}
\usage{
scoring
}
\description{
See \code{vignette("scoring_settings")} on how to define custom scoring settings.
}
\keyword{datasets}
