% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_rules.R
\docType{data}
\name{scoring_empty}
\alias{scoring_empty}
\title{Object with potential scoring combinations based on the \code{scoring} object}
\format{
An object of class \code{list} of length 8.
}
\usage{
scoring_empty
}
\description{
Object with potential scoring combinations based on the \code{scoring} object
}
\keyword{datasets}
