% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagueScoring.R
\docType{data}
\name{scoringRules}
\alias{scoringRules}
\title{Default scoring rules}
\format{An object of class \code{list} of length 10.}
\usage{
scoringRules
}
\description{
Default scoring rules
}
\keyword{datasets}

