% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_be_deprecated.R
\name{scrape_source}
\alias{scrape_source}
\title{Scrape data for a specific position from a single source}
\usage{
scrape_source(src, season, week, position)
}
\description{
Scrape data for a specific position from a single source
}
