% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{set_tiers}
\alias{set_tiers}
\title{Determine Tiers by Position}
\usage{
set_tiers(data_tbl, d_threshold = NULL, src_points)
}
\arguments{
\item{data_tbl}{An output from the \link{projected_points} function}

\item{d_threshold}{THe thresholds to use when applying Cohens'd D function to
determine the tiers. If omitted then the \link{default_threshold} will be used.}

\item{src_points}{An output from the \link{source_points} function}
}
\description{
This function determines tiers for each position by applying Cohen's D effect
size
}
