% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites}
\alias{sites}
\title{Site data}
\format{A data.table with 17 rows and 5 columns
\describe{
 \item{siteId}{Unique integer identifier for the site}
 \item{siteName}{The name of the site}
 \item{subscription}{Indicator whether the site requires subscription to get
 to the data}
 \item{playerId}{Name of column in the player data table that holds the id
 for the players}
}}
\usage{
sites
}
\description{
Data table with information on sites that are being used for datascrapes.
}
\keyword{datasets}

