% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siteTables}
\alias{siteTables}
\title{Table information}
\format{A \link[data.table]{data.table} with 111 rows and 9 columns.
\describe{
 \item{tableId}{An integer that uniquely identifies the table}
 \item{position}{Name of the position that the table holds data for}
 \item{positionAlias}{The position identifier used in the URL for the table}
 \item{siteId}{Integer identfying the site that the table is found on. See
 \link{sites} for values}
 \item{startPage}{If the table spans more pages then the start number for the
 pages. If all the data is on one page the value is 1.}
 \item{endPage}{If the table spans more pages then the end number for the page
 sequence. If all the data is on one page the value is 1}
 \item{stepPage}{If the table spans more page then the step number of the page
 sequence. If all the data is on one page the value is 1}
 \item{season}{Indicates whether the table can be used for seasonal data scrapes.}
 \item{weekly}{Indicates whether the table can be used for weekly data scrapes.}
}}
\usage{
siteTables
}
\description{
Data table with information on tables that data will be scraped from.
}
\keyword{datasets}

