% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siteUrls}
\alias{siteUrls}
\title{URL information}
\format{A \link[data.table]{data.table} with 31 rows and 7 columns.
\describe{
 \item{siteId}{An integer identifying the site the URL is referring to. See
 \link{sites} for values.}
 \item{siteUrl}{The URL that the data will be scraped from. Use placeholders
 for parameters that needs to be substituted, like position, analyst, page,
 season and week.}
 \item{urlPeriod}{Specifies whether the URL is used for week or season data.}
 \item{urlType}{Specifies the data type that is returned from the URL. Could be
 HTML, XML, csv or file}
 \item{urlTable}{The table number that the data scrape should read from the
 URL. Mostly used for HTML data, but can also be used to designate sheets in
 a spreadsheet file.}
 \item{playerLink}{A string to identify the links to player profile pages on
 an HTML page.}
}}
\usage{
siteUrls
}
\description{
Data table with information on the URLs that data will be scraped from.
}
\keyword{datasets}

