% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{source_points}
\alias{source_points}
\title{Calculate Source Points}
\usage{
source_points(data_result, scoring_rules)
}
\arguments{
\item{data_result}{An output from the \link{scrape_data} function.}

\item{scoring_rules}{The scoring rules to be used.}
}
\description{
Function to calculate the projected points for each source.
}
