% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceAnalyst.R
\docType{class}
\name{sourceAnalyst-class}
\alias{sourceAnalyst}
\alias{sourceAnalyst-class}
\title{Class to represent the source Analysts}
\description{
Class to represent the source Analysts
}
\section{Slots}{

\describe{
\item{\code{analystName}}{The name of the analysts}

\item{\code{sourceId}}{The id for the source. This is only used if there are multiple
analysts for a site}

\item{\code{analystId}}{A character string specifying the id to be used for the analyst.
If left blank, it will be set as a 4 letter abbreviation of the site + analyst
names using \link{abbreviate}.}
}}
\examples{
cbs.avg <- sourceAnalyst(analystName = "Average",
                         sourceId = "avg",
                         analystId = "cbav")
}

