% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSite.R
\docType{class}
\name{sourceSite-class}
\alias{sourceSite}
\alias{sourceSite-class}
\title{Class to represent source sites}
\description{
Source sites are one of the foundations to the data scrapes. They are
representing the web sites that provides data for the projections. The
sources can be different for weekly and seasonal data and the type of
data can be different. If the data is scraped from HTML then the playerId
can be derived as well.
}
\section{Slots}{

\describe{
\item{\code{siteId}}{The ID for the site as identfied in configuration data}

\item{\code{siteName}}{Name of source site}

\item{\code{siteUrl}}{String that represents the URL that the site uses for the seasonal data, if any.}

\item{\code{urlType}}{String that identifies the type of data in  seasonURL (HTML, XML, CSV or file)}

\item{\code{urlTable}}{A character string identifying which table to grab from seasonURL}

\item{\code{playerLink}}{A string representing part of the URL to a player profile}

\item{\code{playerId}}{What to call the id number for player if present}
}}
\note{
When specifying the URLs paramters can be used as place holders.
}
\examples{
sourceSite(siteName = "CBS",
           siteUrl = "http://www.cbssports.com/fantasy/football/stats/weeklyprojections/{$Pos}/season/{$SrcID}/standard?&print_rows=9999",
           urlType = "html",
           urlTable = "1",
           playerLink = "/fantasyfootball/players/playerpage/[0-9]{3,6}",
           playerId = "cbsId")
}

