% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceTable.R
\docType{class}
\name{sourceTable-class}
\alias{sourceTable}
\alias{sourceTable-class}
\title{Class to represent a data table from a source site}
\description{
Source table class extends the \link{sourceAnalyst} class
}
\section{Slots}{

\describe{
\item{\code{sourcePosition}}{The position designation that the table represents}

\item{\code{positionAlias}}{The designation that the source site uses for the position}

\item{\code{startPage}}{If the table covers multiple pages, the start numbering for the
pages. Otherwise 1}

\item{\code{endPage}}{If the table covers multiple pages, the end numbering for the
pages. Otherwise 1}

\item{\code{stepPage}}{If the table covers multiple pages, the step in the sequence of
page numbering. Otherwise 1}

\item{\code{tableId}}{ID from configuration data that identifies the table uniquely}
}}
\examples{
cbs.qb <- sourceTable(sourcePosition = "QB",
                      positionAlias = "QB",
                      startPage = 1,
                      endPage = 1,
                      stepPage = 1,
                      tableId = 1
                     )
}

