% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_aggregation.R
\name{stats_by_category}
\alias{stats_by_category}
\title{Organize scraped data by stat category}
\usage{
stats_by_category(data_results)
}
\description{
Breaks the data into the following categories, and imputes missing values
based on rate stats across sources. For example, if passing attemts are missing
then they will be imputed by calculating pass yards per attempt across sources
and dividing the pass yards for the source with pass yards per attempt across
sources.
\describe{
  \item{pass}{Passing stats}
  \item{rush}{Rushing stats}
  \item{rec}{Receiving stats}
  \item{misc}{Miscellaneous offensive stats}
  \item{ret}{Return stats}
  \item{kick}{Kicking stats}
  \item{dst}{DST stats}
  \item{idp}{IDP stats}
}
}
