% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnVariables.R, R/data.R
\docType{data}
\name{tableColumns}
\alias{resultColumns}
\alias{staticColumns}
\alias{tableColumns}
\title{Constants defining result columns}
\format{A \link[data.table]{data.table} with 2924 rows and 6 columns.
\describe{
 \item{tableId}{An integer identifying the table that the column belongs to.
 See \link{siteTables} for values}
 \item{columnName}{The name of the column}
 \item{columnType}{The data type for the column}
 \item{columnOrder}{The order in which the column appears in the table}
 \item{columnPeriod}{Indicates if the column appears in tables used for seasonal
 or weekly data scrapes}
 \item{removeColumn}{Indicates if the column can be removed before the table
 is returned.}
}}
\usage{
staticColumns

resultColumns

tableColumns
}
\description{
The results in the \linkS4class{dataResult} table are determined by the values in
these constants. See definitiions in details.
\describe{
 \item{staticColumns}{Names of columns in all result tables.}
 \item{resultColums}{A list of character vectors (one per position), indicating
the columns in the results table for that position.}
}

Data table with information on the columns in tables identified in \link{siteTables}.
}
\keyword{datasets}

