% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tableRowRemove}
\alias{tableRowRemove}
\title{Table rows to remove}
\format{A \link[data.table]{data.table} with 28 rows and 2 columns.
\describe{
 \item{tableId}{An integer identifying the table that the row can be removed from.
 See \link{siteTables} for values}
 \item{rowRemove}{An integer identifying the row that can be removed}
}}
\usage{
tableRowRemove
}
\description{
Data table with information on rows that will need to be removed from \link{siteTables}
for the data scrape to be succesfull
}
\keyword{datasets}

