% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagueScoring.R
\docType{data}
\name{tierGroups}
\alias{tierGroups}
\title{Default number of tiers for clusters}
\format{An object of class \code{numeric} of length 9.}
\usage{
tierGroups
}
\description{
Default number of tiers for clusters
}
\keyword{datasets}

