% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagueScoring.R
\docType{data}
\name{vorAdjustment}
\alias{vorAdjustment}
\title{Default VOR Adjustments}
\format{An object of class \code{numeric} of length 6.}
\usage{
vorAdjustment
}
\description{
Default VOR Adjustments
}
\keyword{datasets}

