% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagueScoring.R
\docType{data}
\name{vorBaseline}
\alias{vorBaseline}
\title{Default VOR Baseline}
\format{An object of class \code{numeric} of length 6.}
\usage{
vorBaseline
}
\description{
Default VOR Baseline
}
\keyword{datasets}

