% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{weighted.sd}
\alias{weighted.sd}
\title{Calculate Weighted Standard Deviation}
\usage{
weighted.sd(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{The observations to calculate the standard deviations from}

\item{w}{The weights associated with each observation.}

\item{na.rm}{If \code{TRUE}, then NA values will be removed.}
}
\description{
Function to calculate weighted standard deviation.
}
