% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{whdquantile}
\alias{whdquantile}
\title{Weighted Harrell-Davis quantile estimator}
\usage{
whdquantile(x, w = NA, probs, na.rm)
}
\description{
From: https://aakinshin.net/posts/weighted-quantiles/
}
