% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\name{wtd_q}
\alias{wtd_q}
\title{Temp weighted quantile function}
\usage{
wtd_q(x, w, probs, na.rm = FALSE)
}
\description{
From: https://stats.stackexchange.com/questions/373269/why-does-this-simple-weighted-quantile-differ-from-hmiscwtd-quantile-which-me
}
