% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_classes.R
\docType{class}
\name{xlsx_source}
\alias{xlsx_source}
\title{xlsx source object}
\format{An R6 class object.}
\usage{
xlsx_source
}
\description{
This objects represent the basic set of data needed for defining a projection
source with xlsx data. It is an extesnsion of the \link{projection_source}
object with some specific fields and methods related to scraping xlsx data.
}
\section{Methods}{

\describe{
  \item{\code{scrape}}{Scrape the data from the source based on season, week
  and position provided}
}
}

\seealso{
Other source-class: \code{\link{html_source}},
  \code{\link{json_source}},
  \code{\link{projection_source}}
}
\concept{source-class}
\keyword{internal}
