% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projections.R
\docType{data}
\name{default_weights}
\alias{default_weights}
\title{Default Weights for Projection Sources}
\format{
An object of class \code{numeric} of length 14.
}
\usage{
default_weights
}
\description{
These are the weights that are used for each source when calculation weighted
averages and standard deviations if no weights are specified.
\code{c(CBS = 0.344, Yahoo = 0.400,  ESPN = 0.329,  NFL = 0.329,
FFToday = 0.379, NumberFire = 0.322, FantasyPros = 0.000,
FantasySharks = 0.327, FantasyFootballNerd = 0.000,
Walterfootball = 0.281, RTSports = 0.330,
FantasyData = 0.428, Fleaflicker = 0.428, FanDuel = 0.322)}
}
\keyword{datasets}
