% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp_functions.R
\name{espn_draft}
\alias{espn_draft}
\title{Get ADP/AAV data from ESPN}
\usage{
espn_draft(metric = c("adp", "aav"))
}
\arguments{
\item{metric}{Indicates whether APD or AAV data is scraped. By default, ADP
data is scraped. Set it to \code{"aav"} to scrape AAV data.}
}
\value{
A \link{data.frame} with the results.
}
\description{
This function scrapes ADP or AAV data from ESPN
}
