% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjections.R
\name{getProjections}
\alias{getProjections}
\title{Calculate Projected Points}
\usage{
getProjections(scrapeData = NULL, avgMethod = "average",
  leagueScoring = scoringRules, teams = 12, format = "standard",
  mflMocks = NULL, mflLeagues = NULL, adpSources = c("CBS", "ESPN", "FFC",
  "MFL", "NFL"))
}
\arguments{
\item{scrapeData}{The scraped projections data from \link{runScrape}.}

\item{avgMethod}{Which average method should be used for aggregating the
projections from different sources.}

\item{leagueScoring}{List of scoring rules for the league see \link{scoringRules}
for an example}

\item{teams}{Number of teams in the league (integer)}

\item{format}{League format}

\item{mflMocks}{Include mock drafts from MFL. Set to 1 if only mock drafts
should be used, 0 if only real drafts should be used. If not speficied all
types of drafts will be used.}

\item{mflLeagues}{What type of leagues to include for MFL. Set to 0 to use
redraft leagues only; 1 to only use keeper leagues, 2 for rookie drafts, and
3 for MFL Public Leagues. If not speficied all types of drafts will be used.}

\item{ADPsource}{Character vector with one or more of \code{c("CBS", "ESPN", "FFC", "MFL", "NFL")}}
}
\description{
Calculate projected fantasy points, confidence intervals, risk, tiers, etc.
}

