% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjections.R
\name{getProjections}
\alias{getProjections}
\title{Calculate Projected Points}
\usage{
getProjections(scrapeData = NULL, avgMethod = "average",
  leagueScoring = scoringRules, vorBaseline, vorType, teams = 12,
  format = "standard", mflMocks = NULL, mflLeagues = NULL,
  adpSources = c("CBS", "ESPN", "FFC", "MFL", "NFL"), getADP = TRUE,
  getECR = TRUE, writeFile = TRUE)
}
\arguments{
\item{scrapeData}{The scraped projections data from \link{runScrape}.}

\item{avgMethod}{A string specifying which average method to use for aggregating the
projections from different sources: mean ("average"), robust average ("robust"), or weighted average ("weighted"). Defaults to mean. Edit the analysts' weights for the weighted average in the \link{analysts} table.}

\item{leagueScoring}{List of scoring rules for the league see \link{scoringRules}
for an example.}

\item{vorBaseline}{The numbers (position rank values or point values) at each position to use for the baseline when
calculating VOR.}

\item{vorType}{Whether the baseline numbers are ranks or points. Defaults to position ranks.}

\item{teams}{Number of teams in the league (integer).}

\item{format}{League format ("standard" for standard leagues or "ppr" for Point-Per-Reception leagues).}

\item{mflMocks}{Whether to include mock drafts from MyFantasyLeague.com (MFL). Set to 1 to use only mock drafts,
0 to use only real drafts. If not specified, all draft types will be used.}

\item{mflLeagues}{What type of leagues to include for MyFantasyLeague.com (MFL). Set to 0 to use
only redraft leagues; 1 to use only keeper leagues, 2 for rookie drafts, and
3 for MFL Public Leagues. If not specified, all draft types will be used.}

\item{ADPsource}{Character vector with one or more of \code{c("CBS", "ESPN", "FFC", "MFL", "NFL")}.}
}
\description{
Calculate projected fantasy points, confidence intervals, risk, tiers, etc.
}
\examples{
getProjections(scrapeData,                    ## Based on data in scrapeData
               avgMethod = "weighted",        ## calculate the projections using a weighted average
               leagueScoring = scoringRules,  ## using defined scoringRules,
               vorBaseline, vorType,          ## VOR Baselines and types
               teams = 12, format = "ppr",    ## for a 12 team ppr league
               mflMocks = 0, mflLeagues = 0,  ## using only real MFL redraft league
               adpSources =  c("FFC", "MFL")) ## and ADP data from MFL and FFC
}

