% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_classes.R
\docType{class}
\name{json_source}
\alias{json_source}
\title{JSON source object}
\format{An R6 class object.}
\usage{
json_source
}
\description{
This objects represent the basic set of data needed for defining a projection
source with JSON data. It is an extesnsion of the \link{projection_source}
object with some specific fields and methods related to scraping JSON data.
}
\section{Fields}{

\describe{
\item{\code{json_elem}}{String containing the name of the JSON element in the result
that identifies the data}

\item{\code{stat_elem}}{String containing the name of the stats element in the JSON
result if applicable}

\item{\code{player_elem}}{Character vector containing the name of the player elements
in the JSON result if applicable}

\item{\code{player_cols}}{Named character vector used to rename the player columns in
the JSON result if needed.}
}}

\section{Methods}{

\describe{
  \item{\code{scrape}}{Scrape the data from the source based on season, week
  and position provided}
}
}

\seealso{
Other source-class: \code{\link{html_source}},
  \code{\link{projection_source}},
  \code{\link{xlsx_source}}
}
\keyword{internal}
