% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScrape.R
\name{runScrape}
\alias{runScrape}
\title{Run a data scrape}
\usage{
runScrape(week = NULL, season = NULL, analysts = NULL, positions = NULL)
}
\arguments{
\item{week}{The week number that the scrape is going to be executed for}

\item{season}{The season that the scrape will be executed for}

\item{analysts}{An integer vector of analystIds specifying which analysts to
execute the scrape for. See \link{analysts} data set for values.}

\item{positions}{A character vector of position names specifying which positions
to execute the scrape for}
}
\value{
list of \link{dataResults}. One entry per position scraped.
}
\description{
Executing a data scrape based on inputs of week, season, analysts and positions.
If no inputs are specified the user will be prompted.
}
\note{
Historical scrapes are nearly impossible to do as very few if any sites
actually stores their historical projections. An attempt to scrape historical
projections will likely produce current projectsions in most cases.
}

