% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp_functions.R
\name{yahoo_draft}
\alias{yahoo_draft}
\title{Get ADP/AAV data from Yahoo}
\usage{
yahoo_draft(metric = c("adp", "aav"))
}
\arguments{
\item{metric}{Indicates whether ADP or AAV data is scraped. By default, ADP
data is scraped. Set it to \code{"aav"} to scrape AAV data.}
}
\value{
A \link{data.frame} with the results.
}
\description{
This function scrapes ADP or AAV data from Yahoo Sports
}
